+++ src/main/java/org/GehtSoftHWByAziz/HW1/CustomList.java
@@ -1,4 +1,4 @@
-package org.GehtSoftHWByAziz;
+package org.GehtSoftHWByAziz.HW1;
 
 import java.util.*;
